test TAP device only on libev-badvpn tunnel; rename go vpn tunnel file; docs
