--- src/electron/tap.ts
@@ -1,57 +1,0 @@
-// Copyright 2021 The Outline Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import {execSync} from 'child_process';
-
-import * as errors from '../www/model/errors';
-
-// Raises an error if:
-//  - the TAP device does not exist
-//  - the TAP device does not have the expected IP/subnet
-//
-// Note that this will *also* throw if netsh is not on the PATH. If that's the case then the
-// installer should have failed, too.
-//
-// Only works on Windows!
-export function testTapDevice(tapDeviceName: string, tapDeviceIp: string) {
-  // Sample output:
-  // =============
-  // $ netsh interface ipv4 dump
-  // # ----------------------------------
-  // # IPv4 Configuration
-  // # ----------------------------------
-  // pushd interface ipv4
-  //
-  // reset
-  // set global icmpredirects=disabled
-  // set interface interface="Ethernet" forwarding=enabled advertise=enabled nud=enabled
-  // ignoredefaultroutes=disabled set interface interface="outline-tap0" forwarding=enabled
-  // advertise=enabled nud=enabled ignoredefaultroutes=disabled add address name="outline-tap0"
-  // address=10.0.85.2 mask=255.255.255.0
-  //
-  // popd
-  // # End of IPv4 configuration
-  const lines = execSync(`netsh interface ipv4 dump`).toString().split('\n');
-
-  // Find lines containing the TAP device name.
-  const tapLines = lines.filter(s => s.indexOf(tapDeviceName) !== -1);
-  if (tapLines.length < 1) {
-    throw new errors.SystemConfigurationException(`TAP device not found`);
-  }
-
-  // Within those lines, search for the expected IP.
-  if (tapLines.filter(s => s.indexOf(tapDeviceIp) !== -1).length < 1) {
-    throw new errors.SystemConfigurationException(`TAP device has wrong IP`);
-  }
-}
